+++ src/async_impl.rs
@@ -109,14 +109,6 @@ impl YahooConnector {
         Ok(YSearchResult::from_opt(&result))
     }
 
-    /// Get list for options for a given name
-    pub async fn search_options(&self, name: &str) -> Result<YOptionChain, YahooError> {
-        let url = format!("https://query2.finance.yahoo.com/v6/finance/options/{name}");
-        let resp = self.client.get(url).send().await?.text().await?;
-
-        Ok(serde_json::from_str(&resp)?)
-    }
-
     // Get symbol metadata
     pub async fn get_ticker_info(&mut self, symbol: &str) -> Result<YQuoteSummary, YahooError> {
         if let None = &self.crumb {
@@ -487,14 +479,6 @@ mod tests {
         assert_eq!(&result[0].meta.data_granularity, "1d");
         let capital_gains = response.capital_gains().unwrap();
         assert!(capital_gains.len() > 0usize);
-    }
-
-    #[test]
-    fn search_options() {
-        let provider = YahooConnector::new().unwrap();
-        let resp = tokio_test::block_on(provider.search_options("AAPL"));
-
-        assert!(resp.is_ok());
     }
 
     #[test]